+++ src/pages/A/Join.tsx
