+++ src/pages/Z/SplashScreen.tsx
